fix: Read from filepath instead of using open file
Some editors like Vim and Nova move the original file to create a
backup, and then create a second file in its place for editing.
This meant the original file didn't have changes from the edit
session, and couldn't read the entered title. This commit reads in
text from the original filepath instead of the open file that was
moved, and thus can suggest a title.